Move deposit action logic to its own folder
